backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "uvicorn>=0.41.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
